Add sub-package for unix and move deb and rpm there.
