+++ src/com/inet/gradle/setup/unix/deb/DebUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.inet.gradle.setup.deb;
+package com.inet.gradle.setup.unix.deb;
 
 import java.io.File;
 import java.io.IOException;
