+++ src/com/inet/gradle/setup/unix/deb/template/postrm.sh
