+++ src/com/inet/gradle/setup/unix/deb/template/prerm.sh
