+++ src/com/inet/gradle/setup/unix/rpm/RpmControlFileBuilder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.inet.gradle.setup.rpm;
+package com.inet.gradle.setup.unix.rpm;
 
 import java.io.File;
 import java.io.FileOutputStream;
