+++ src/com/inet/gradle/setup/unix/rpm/template/init-service.sh
